Some more cleanup & refactor
